all logic from MainWindow moved to App
