+++ TranslatorS3/App.xaml
@@ -2,7 +2,8 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:TranslatorS3"
-             StartupUri="MainWindow.xaml">
+             >
+             <!--StartupUri="MainWindow.xaml">-->
     <Application.Resources>
          
     </Application.Resources>
